+++ LeProHQpy/src/LeProHQ/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-from .cg0 import cg0
-from .cg1 import cg1
-from .cgBar1 import cgBar1, cgBarF1, cgBarR1
-from .cq1 import cq1
-from .cqBarF1 import cqBarF1
-from .dq1 import dq1
+from .cg0_ import cg0
+from .cg1_ import cg1
+from .cgBar1_ import cgBar1, cgBarF1, cgBarR1
+from .cq1_ import cq1
+from .cqBarF1_ import cqBarF1
+from .dq1_ import dq1
